Вынес sqlite в отдельный модуль
